requirements.txt
fastapi==0.111.0
uvicorn==0.40.0
pydantic==2.12.4
pydantic-settings==2.11.0
python-multipart==0.0.9
python-dotenv==1.2.1
requests==2.32.5
httpx==0.28.1
aiohttp==3.9.5
pandas==2.2.3
numpy==1.26.4
yfinance==0.2.66
groq==0.30.0
langchain==0.3.25
langchain-core>=0.3.58,<1.0.0
langchain-groq==0.3.6
tavily-python==0.7.13
fastmcp==2.13.1
mcp==1.22.0
matplotlib==3.9.1
seaborn==0.13.2
plotly==5.24.1
Pillow==10.4.0
xhtml2pdf==0.2.17
jinja2==3.1.4
markdown==3.10
reportlab==4.4.3
beautifulsoup4==4.13.3
feedparser==6.0.12
python-dateutil==2.9.0.post0
scipy==1.13.1
pyyaml==6.0.2
sse-starlette==3.0.3
assemblyai==0.48.4
faster-whisper==1.1.1
sounddevice==0.5.1
pyaudio==0.2.14